@jtgeibel 2019-03-20
Looks great to me!  It took me way too long to notice that the pagination helper was adding the total count for the original queries.

@bors r+